+++ R/render.R
@@ -116,7 +116,7 @@ moon_reader = function(
     },
     tags$script(HTML(paste(c(sprintf(
       'var slideshow = remark.create(%s);', if (length(nature)) xfun::tojson(nature) else ''
-    ), pkg_file(c('js/show-widgets.js', 'js/print-css.js', 'js/after.js')),
+    ), pkg_file(c('js/show-widgets.js', 'js/print-css.js', 'js/after.js', 'js/script-tags.js')),
     play_js, countdown_js, hl_pre_js), collapse = '\n')))
   )), tmp_js)